requirements.txt
fastapi
uvicorn
groq
rank-bm25
python-dotenv
